+++ nutree/dot.py
@@ -7,7 +7,7 @@ Functions and declarations to support
 from __future__ import annotations
 
 from pathlib import Path
-from typing import IO, TYPE_CHECKING, Iterator, Union
+from typing import IO, TYPE_CHECKING, Iterator, Optional, Union
 
 from .common import MapperCallbackType, call_mapper
 
@@ -16,7 +16,7 @@ if TYPE_CHECKING:  # Imported by type checkers, but prevent circular includes
     from .tree import Tree
 
 try:
-    import pydot
+    import pydot  # type: ignore
 except ImportError:
     pydot = None
 
@@ -117,8 +117,8 @@ def tree_to_dotfile(
     graph_attrs=None,
     node_attrs=None,
     edge_attrs=None,
-    node_mapper: MapperCallbackType = None,
-    edge_mapper: MapperCallbackType = None,
+    node_mapper: Optional[MapperCallbackType] = None,
+    edge_mapper: Optional[MapperCallbackType] = None,
 ) -> None:
     if isinstance(target, str):
         target = Path(target)